Unir iy hacer algunas modificaciones a las partes de las compañeras
